moving utils into lib folder
